Add audio processing example
